requirements.txt
gradio
transformers
PyPDF2
pdfminer.six

